+++ public/images-extension/bouton-notifications-colored.png
