+++ public/images-extension/suggestion-activate.png
