+++ src/interfaces/Anime.js
@@ -1,0 +1,13 @@
+class anime {
+  constructor() {
+    this.name = "";
+    this.title = "";
+    this.episode = "";
+    this.saison = 0;
+    this.link = "";
+    this.duration = 0;
+    this.currentTime = 0;
+    this.notif = false;
+    this.lastUpdate = 0;
+  }
+}
