+++ src/services/StorageService.js
@@ -1,0 +1,23 @@
+/**
+ * StorageService
+ * Encapsule chrome.storage.local pour simplifier les appels.
+ */
+export const StorageService = {
+  async get(key) {
+    return new Promise((resolve) => {
+      chrome.storage.local.get(key, (result) => resolve(result[key]));
+    });
+  },
+
+  async set(key, value) {
+    return new Promise((resolve) => {
+      chrome.storage.local.set({ [key]: value }, resolve);
+    });
+  },
+
+  async remove(key) {
+    return new Promise((resolve) => {
+      chrome.storage.local.remove(key, resolve);
+    });
+  }
+};
