+++ src/vars.js
@@ -1,0 +1,10 @@
+/**
+ * File Name      : config.js
+ * Description    : This file handles user login for the application.
+ * Author         : Mathis Gramage, Mathis Cucherat
+ * Date           : Last update 2025-10-07
+ * Version        : 1.0.0
+ */
+
+export const URL_API = "http://localhost/Ani-Api/api/connexion";
+export const EDITOR_EXTENSION_ID = "olggkeglcmmolkpmnpffffcpcpdlonpk";
